% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_atlas_data.R
\name{generate_atlas_data}
\alias{generate_atlas_data}
\title{Generate Atlas Data}
\usage{
generate_atlas_data(AP, half = TRUE)
}
\arguments{
\item{AP}{(required) \code{numeric} AP in mm from bregma. Will be rounded by \code{\link[=roundAP]{roundAP()}}.}

\item{half}{\code{boolean} whether to plot half or full section. Default is TRUE.}
}
\value{
A \code{data.frame} with path information and style information that can be used to plot the atlas plates.
}
\description{
This function generates the atlas data needed to plot. You can use it to generate your own custom plots.
}
\seealso{
\code{\link[=plot_atlas]{plot_atlas()}}, \code{\link[=roundAP]{roundAP()}}
}
