% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobrainr-package.R
\docType{package}
\name{nobrainr-package}
\alias{nobrainr}
\alias{nobrainr-package}
\title{nobrainr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{matiasandina@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
