% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_atlas.R
\name{plot_atlas}
\alias{plot_atlas}
\title{Plot Atlas}
\usage{
plot_atlas(AP, half = TRUE, bw = FALSE, include_legend = FALSE, verbose = TRUE)
}
\arguments{
\item{AP}{(required) \code{numeric} AP in mm from bregma. Will be rounded by \code{\link[=roundAP]{roundAP()}}.}

\item{half}{\code{boolean} whether to plot half or full section. Default is TRUE.}

\item{bw}{\code{boolean} whether to plot using standard Allen Atlas colors or in black and white. Default is FALSE.}

\item{include_legend}{\code{boolean} whether to include legend with the regions. Default is FALSE.}

\item{verbose}{\code{boolean} whether to give feedback. Default to TRUE.}
}
\description{
This function will plot a desired coronal plate using the desired \code{AP} level in mm from bregma.
}
\seealso{
\code{\link[=roundAP]{roundAP()}}
}
