% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roi_counts.R
\name{plot_roi_counts}
\alias{plot_roi_counts}
\title{Plot ROI counts}
\usage{
plot_roi_counts(count_df, AP, half = TRUE)
}
\arguments{
\item{count_df}{(required) \code{data.frame} with counts per region (see \code{trial_df} for example).}

\item{AP}{(required) \code{numeric} AP in mm from bregma. Will be rounded by \code{\link[=roundAP]{roundAP()}}.}

\item{half}{\code{boolean} whether to plot half or full section. Default is TRUE.}
}
\description{
This function will plot a desired coronal plate using the desired \code{AP} level in mm from bregma. It will color each roi using the counts provided in \code{count_df}.
}
\seealso{
\code{\link[=plot_atlas]{plot_atlas()}}, \code{\link[=roundAP]{roundAP()}}
}
