% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundAP.R
\name{roundAP}
\alias{roundAP}
\title{Round AP}
\usage{
roundAP(AP, plane = "coronal")
}
\arguments{
\item{AP}{\code{numeric} AP coordinate in mm from bregma.}

\item{plane}{\code{string} must be "coronal" in this version of the package. "Sagittal" not yet supported.}
}
\value{
The rounded \code{AP} value to the closest plane in the \code{atlasIndex} dataset.
}
\description{
Round AP
}
